Started implementing multithreading
Got some multithreading working, also implemented more events that are handled.
Need to split the threads into 3 - one waits for xevents and sends them away, second one waits for xevents and transforms them into events and a main thread.

+++ src/platform/linux/x11/atoms.rs
@@ -1,5 +1,5 @@
-type Atom = u64;
-#[derive(Clone, Copy)]
+type Atom = usize;
+#[derive(Debug, Clone)]
 pub struct Atoms {
     pub wm_delete_window: Atom,
 }
